refactor(components): update import paths for utility functions and enhance Tailwind CSS configuration
