+++ frontend-customer/tailwind.config.ts
@@ -7,7 +7,9 @@ export default {
     "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
-	 "./src/views/**/*.{js,ts,jsx,tsx,mdx}",
+	"./src/views/**/*.{js,ts,jsx,tsx,mdx}",
+	'./src/libs/**/*.{js,ts,jsx,tsx,mdx}',
+    'node_modules/antd/dist/antd.css',
   ],
   theme: {
   	extend: {
